options: Add watch-later-use-title
A new option has been added to override the filename in
mp_get_playback_resume_config_filename.

This option was made for the case where the video path or url changes or
isn't unique but the user still wants progress to be saved.

Examples include
- streaming a url where the upstream path contains transient information
like a session id
- streaming a url where irrelevant query parameters are added/removed.
- Playing from stdin since the "file name" is always "-"
- Playing a local file from a link vs the real file

This patch makes the assumption that that the video is actually the same
and will use hash of the media title instead of the hash of the path.
If a title was explicitly specified, its hash will be used instead.